Updated Chapter 5's Blackjack dynamics to correctly handing the situation where the player or dealer receives an ace while already having a usable ace.
It is possible that the new sum requires either one or both aces to act as a 1. In the case where only one ace is treated as a 1, the player or dealer still maintains a single usable ace. This was not handled correctly before (0 usable aces would be present after one of the two aces was treated as a 1).