Update the axis limit
I think this should be better and closer to the original one as the X axis should start from 1 and the Y axis should be limited to 5.
Besides, I don't know why the line 367 and the line 369 were commented, so I uncommented them. I added more description to the Y label too.